move hashing functionality to accounts_hash
