+++ runtime/src/lib.rs
@@ -3,6 +3,7 @@ pub mod accounts;
 pub mod accounts_background_service;
 pub mod accounts_cache;
 pub mod accounts_db;
+pub mod accounts_hash;
 pub mod accounts_index;
 pub mod append_vec;
 pub mod bank;
